preprocess-plans/worker/requirements.txt
google-genai>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
